dags/requirements.txt
# Airflow requirements
requests>=2.28.0
apache-airflow==2.8.0
psycopg2-binary==2.9.7
